aider: Added a link to https://hbar.watch in the README.md file.
# Context:
USER: add to the `README.md` a link to https://hbar.watch so people can get to the site.
ASSISTANT: To add a link to the website in the `README.md` file, you need to modify the file as follows:

```markdown
README.md
<<<<<<< ORIGINAL
# hbar.watch
=======
# [hbar.watch](https://hbar.watch)
>>>>>>> UPDATED
```

This change will make the title of the README a clickable link that directs to https://hbar.watch.